backend/requirements.txt
flask==2.3.0
flask-cors==4.0.0
flask-jwt-extended==4.5.2
flask-sqlalchemy==3.1.0
pandas>=2.2.0
python-dotenv==1.0.0
werkzeug==2.3.0